requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
sqlparse==0.5.1
typing_extensions==4.12.2
djangorestframework==3.15.2
django-rest-swagger==2.2.0
django-environ==0.4.5
psycopg2==2.9.5
django-oauth-toolkit==3.0.1
drf-yasg==1.21.7
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-redis==5.4.0
redis==4.3.4
django-cors-headers==4.4.0